Move azure and credentials package to utils
