--- hack/template/azure.json
@@ -1,9 +1,0 @@
-{
-    "cloud": "{{.Cloud}}",
-    "tenantId": "{{.TenantID}}",
-    "subscriptionId": "{{.SubscriptionID}}",
-    "aadClientId": "{{.AADClientID}}",
-    "aadClientSecret": "{{.AADClientSecret}}",
-    "resourceGroup": "{{.ResourceGroup}}",
-    "location": "{{.Location}}"
-}
